App-Country-life-index-comparisons
This application have four pages and pages contain different apps inside them. The dataset is called 'Quality of life index by countries 2020' and it contains 10 different variables and 80 observations. The first variable is Country. 2nd variable up to 10th variable are representing the countries' specific demographic indexes, such as safety index, purcashing power index, etc. In order to obtain better visualizations, three more columns added; longitude, latitude, and Continent, respectively.
The first app simply compares the two selected demographic indexes and checks the relationship between them with a plot.
The second app compares the selected continents by using the selected demographic index. User may reach general information about the selected continents with a check box. A leaflet map can be seen at the top of the main panel which contains the dataset countries in the selected continent, and with the help of the bar plot, user can see the overall comparison of countries with the selected demographic index.
In the third app, user selects a limit for the quality of life index. Then app shows the countries that satisfies the index in the leaflet map with the pop-ups. User may click the markers to see additional information about the country.
The fourth app offers the user five countries according to the user's preference among the indexes from -5 to 5. With basic calculations, app simply multiplies the index with the selected value (-5 to 5), and then adds or substracts according to the impact of the index.